+++ packages/atproto/lib/src/service/repositories/repositories_service.dart
@@ -14,20 +14,24 @@ abstract class RepositoriesService {
     required String did,
     required String service,
     required core.ClientContext context,
+    final core.GetClient? mockedGetClient,
+    final core.PostClient? mockedPostClient,
   }) =>
       _RepositoriesService(
         did: did,
         service: service,
         context: context,
+        mockedGetClient: mockedGetClient,
+        mockedPostClient: mockedPostClient,
       );
 
-  Future<core.ATProtoResponse<Record>> createRecord({
-    required String collection,
+  Future<core.XRPCResponse<Record>> createRecord({
+    required core.NSID collection,
     required Map<String, dynamic> record,
   });
 
-  Future<core.ATProtoResponse<core.Empty>> deleteRecord({
-    required String collection,
+  Future<core.XRPCResponse<core.EmptyData>> deleteRecord({
+    required core.NSID collection,
     required String uri,
   });
 }
@@ -39,38 +43,36 @@ class _RepositoriesService extends ATProtoBaseService
     required super.did,
     required super.service,
     required super.context,
-  });
+    super.mockedGetClient,
+    super.mockedPostClient,
+  }) : super(methodAuthority: 'repo.atproto.com');
 
   @override
-  Future<core.ATProtoResponse<Record>> createRecord({
-    required String collection,
+  Future<core.XRPCResponse<Record>> createRecord({
+    required core.NSID collection,
     required Map<String, dynamic> record,
   }) async =>
-      super.transformSingleDataResponse(
-        await super.post(
-          '/xrpc/com.atproto.repo.createRecord',
-          body: {
-            'did': did,
-            'collection': collection,
-            'record': record,
-          },
-        ),
-        dataBuilder: Record.fromJson,
+      await super.post(
+        'createRecord',
+        body: {
+          'did': did,
+          'collection': collection.toString(),
+          'record': record,
+        },
+        to: Record.fromJson,
       );
 
   @override
-  Future<core.ATProtoResponse<core.Empty>> deleteRecord({
-    required String collection,
+  Future<core.XRPCResponse<core.EmptyData>> deleteRecord({
+    required core.NSID collection,
     required String uri,
   }) async =>
-      super.transformEmptyDataResponse(
-        await super.post(
-          '/xrpc/com.atproto.repo.deleteRecord',
-          body: {
-            'did': did,
-            'collection': collection,
-            'rkey': uri.split('/').last,
-          },
-        ),
+      await super.post<core.EmptyData>(
+        'deleteRecord',
+        body: {
+          'did': did,
+          'collection': collection.toString(),
+          'rkey': uri.split('/').last,
+        },
       );
 }