+++ packages/atproto/test/src/service/repositories/repositories_service_test.dart
@@ -7,7 +7,7 @@ import 'package:atproto/src/service/repositories/repositories_service.dart';
 import 'package:atproto_core/atproto_core.dart' as core;
 import 'package:test/test.dart';
 
-import '../../../mocks/client_context_stubs.dart' as context;
+import '../../../mocks/mocked_clients.dart';
 import '../common_expectations.dart';
 
 void main() {
@@ -16,19 +16,21 @@ void main() {
       final repositories = RepositoriesService(
         did: 'test',
         service: 'test',
-        context: context.buildPostStub(
-          'test',
-          '/xrpc/com.atproto.repo.createRecord',
+        context: core.ClientContext(
+          accessJwt: '1234',
+          timeout: Duration.zero,
+        ),
+        mockedPostClient: createMockedPostClient(
           'test/src/service/repositories/data/create_record.json',
         ),
       );
 
       final response = await repositories.createRecord(
-        collection: 'test.post',
+        collection: core.NSID.create('com.atproto', 'test'),
         record: {},
       );
 
-      expect(response, isA<core.ATProtoResponse>());
+      expect(response, isA<core.XRPCResponse>());
       expect(response.data, isA<Record>());
     });
 
@@ -36,17 +38,19 @@ void main() {
       final repositories = RepositoriesService(
         did: 'test',
         service: 'test',
-        context: context.buildPostStub(
-          'test',
-          '/xrpc/com.atproto.repo.createRecord',
-          'test/src/service/repositories/data/create_record.json',
+        context: core.ClientContext(
+          accessJwt: '1234',
+          timeout: Duration.zero,
+        ),
+        mockedPostClient: createMockedPostClient(
+          'test/src/service/data/error.json',
           statusCode: 401,
         ),
       );
 
       expectUnauthorizedException(
         () async => await repositories.createRecord(
-          collection: 'test.post',
+          collection: core.NSID.create('com.atproto', 'test'),
           record: {},
         ),
       );
@@ -56,17 +60,19 @@ void main() {
       final repositories = RepositoriesService(
         did: 'test',
         service: 'test',
-        context: context.buildPostStub(
-          'test',
-          '/xrpc/com.atproto.repo.createRecord',
-          'test/src/service/repositories/data/create_record.json',
+        context: core.ClientContext(
+          accessJwt: '1234',
+          timeout: Duration.zero,
+        ),
+        mockedPostClient: createMockedPostClient(
+          'test/src/service/data/error.json',
           statusCode: 429,
         ),
       );
 
       expectRateLimitExceededException(
         () async => await repositories.createRecord(
-          collection: 'test.post',
+          collection: core.NSID.create('com.atproto', 'test'),
           record: {},
         ),
       );
@@ -78,37 +84,41 @@ void main() {
       final repositories = RepositoriesService(
         did: 'test',
         service: 'test',
-        context: context.buildPostStub(
-          'test',
-          '/xrpc/com.atproto.repo.deleteRecord',
+        context: core.ClientContext(
+          accessJwt: '1234',
+          timeout: Duration.zero,
+        ),
+        mockedPostClient: createMockedPostClient(
           'test/src/service/repositories/data/delete_record.json',
         ),
       );
 
       final response = await repositories.deleteRecord(
-        collection: 'test.post',
+        collection: core.NSID.create('com.atproto', 'test'),
         uri: '',
       );
 
-      expect(response, isA<core.ATProtoResponse>());
-      expect(response.data, isA<core.Empty>());
+      expect(response, isA<core.XRPCResponse>());
+      expect(response.data, isA<core.EmptyData>());
     });
 
     test('when unauthorized', () async {
       final repositories = RepositoriesService(
         did: 'test',
         service: 'test',
-        context: context.buildPostStub(
-          'test',
-          '/xrpc/com.atproto.repo.deleteRecord',
-          'test/src/service/repositories/data/delete_record.json',
+        context: core.ClientContext(
+          accessJwt: '1234',
+          timeout: Duration.zero,
+        ),
+        mockedPostClient: createMockedPostClient(
+          'test/src/service/data/error.json',
           statusCode: 401,
         ),
       );
 
       expectUnauthorizedException(
         () async => await repositories.deleteRecord(
-          collection: 'test.post',
+          collection: core.NSID.create('com.atproto', 'test'),
           uri: '',
         ),
       );
@@ -118,17 +128,19 @@ void main() {
       final repositories = RepositoriesService(
         did: 'test',
         service: 'test',
-        context: context.buildPostStub(
-          'test',
-          '/xrpc/com.atproto.repo.deleteRecord',
-          'test/src/service/repositories/data/delete_record.json',
+        context: core.ClientContext(
+          accessJwt: '1234',
+          timeout: Duration.zero,
+        ),
+        mockedPostClient: createMockedPostClient(
+          'test/src/service/data/error.json',
           statusCode: 429,
         ),
       );
 
       expectRateLimitExceededException(
         () async => await repositories.deleteRecord(
-          collection: 'test.post',
+          collection: core.NSID.create('com.atproto', 'test'),
           uri: '',
         ),
       );