+++ packages/bluesky/lib/src/service/bluesky_base_service.dart
@@ -12,6 +12,9 @@ abstract class BlueskyBaseService extends core.BaseService {
     required this.atproto,
     required super.service,
     required super.context,
+    required super.methodAuthority,
+    super.mockedGetClient,
+    super.mockedPostClient,
   });
 
   /// The client of AT Protocol.