+++ packages/xrpc/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release Note
 
+## v0.0.3
+
+- Fixed internal structures.
+
+## v0.0.2
+
+- `service` and `getClient`/`postClient` are now nullable. ([#146](https://github.com/myConsciousness/atproto.dart/issues/146))
+
 ## v0.0.1
 
 - First Release.